requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.1
streamlit==1.29.0
openai==1.42.0
requests==2.31.0
sentence-transformers==2.6.1
pypdf==4.2.0
python-dotenv==1.1.1
faiss-cpu==1.7.4
numpy==1.24.3
